dev-requirements.txt
-r requirements.txt

pytest
pytest-django
factory_boy>=3.0.0
pytest-factoryboy
flake8
ipdb
tox